+++ mwclient/_types.py
